Migrate `whatdoineed` to new application
